backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
channels[daphne]
psycopg2-binary
celery
redis
django-cors-headers
